Update patch set 2
Patch Set 2:

Rohan - I think that's unrealistic, to be honest.

It's quite likely there will come a time when you want to break the API. We would just like to be able to predict how the version number will reflect that.

e.g. for oslo.config and oslo.messaging, we will go from 1.x to 2.x if (and only if) we need to break the API

Honestly, any of these are fine approaches:

 python-seamicroclient>=0.1.0,<0.2 # you plan to break the API during the 0.x series, but may release compatible 0.1.x releases
python-seamicroclient>=0.1.0,<1 # you plan to maintain API compat until you release 1.0 with breaking changes
python-seamicroclient>=0.1.0,<2 # you plan to maintain API compat even in 1.0, but will move to 2.x if you make API changes

But I don't think this approach is sensible:

 python-seamicroclient>=0.1.0 # you *never* plan to break the API

Patch-set: 2