Update patch set 2
Patch Set 2:

Nadya Privalova, I can add any dependency that you want in Debian, and if that's in a Python module, it's likely to be sync from my package in Ubuntu. However, downgrading a package is always annoying, and I would prefer to have a patch for the package if possible. Best would anyway to have things fixed upstream, whatever happens.

Patch-set: 2