Six: bump to 1.5.2
In order to ensure Python 3 compatibility of the OpenStack Python clients, we
need more advanced features of six that are not included in six 1.4.1. For
instance, six.moves.xmlrpc_client only landed in six in the 1.4.2 release. Just
bump to the latest release, so that we have as many features as we can dream
of.

Change-Id: I148a9736d8271575647bfd0963771a177f8b6c1e