Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

In theory Thomas is probably right that *if* the upstream project was chronically unresponsive, then we would have no options other than (a) to fork, or (b) to find another equivalent dependency.

However, IMO it's completely unrealistic to hold up all development on the ceilometer hbase driver by requiring that we fork it right now.

The necessary fix may be forthcoming upstream, just slow to land and be released ... in that case, we should have the option to upper-bound the dependency as a pragmatic workaround.

Given the apparent responsiveness of upstream to working on the fix, seems to me the most reasonable option is to wait for the fix while working-around the problem with a version cap in the meantime.

Though if it turns out that months pass and still no fix is released by upstream, we would have to reconsider that strategy.

Cheers,
Eoghan

Patch-set: 3
Reviewer: Gerrit User 2284 <2284@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1