Add Docker requirement
Use the docker-py module instead of custom REST code in Nova.

This reduces the amount of code we need to maintain and test
as part of OpenStack and separates the concern of maintaining
compatibility with upcoming API versions.

(This patch replaces https://review.openstack.org/#/c/72691/1 as
I cannot restore change and update it)

Change-Id: I30cf4f3e2483de1f3081944faf184af8170a31c3