Update patch set 1
Patch Set 1:

I'd argue that even if it gets fixed, it is worth excluding the known-broken version if only because of caches and outdated mirrors.

It isn't clear to me who really owns this package. I guess xcat / Jarrod Johnson? I've sent him an email.

Devananda seems somewhat involved as well and this obviously affects Ironic, so he has already been added as a reviewer.

Patch-set: 1