Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

Given the review criteria:

https://wiki.openstack.org/wiki/Requirements#Review_Criteria

It's apache 2, it's pretty new (December of 2013):

https://github.com/vmware/pyvmomi

Given how new it is I'm assuming it's not packaged in any distros.  It doesn't show up in rpmfind.net at least.

I have no idea if it's py3 compatible.

According to the review criteria, there should be an upper bounds check on major version changes, i.e. < 6, but that depends on the API compat policy in the project (which I don't see documented).

The commit message should also call out which projects are going to be using this, and if there is/are blueprint(s) that need this you should also call this out, i.e. vmware driver refactoring in nova in juno.

Patch-set: 2
Reviewer: Gerrit User 6873 <6873@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1