Update requests to 2.10 and add httmock to tests
In testing REST client code in Neutron, I was seeing the wrong
exception generated, when there was an HTTP timeout on request.
Googled and found that 2.1.0 resolves the issue (and verified
that it did).

To make unit tests easier, would like to include httmock package
from PyPI. This is a mocking library for the requests package
and makes tests much simpler.

Change-Id: I42df522eb841f0798f742e5ff375fcd94908f93d
Closes-Bug: 1282855