Update patch set 2
Patch Set 2:

mark, Sergey,

It seems to be failing one test, and seems to be with the six package and not the two that I changed.

On the previous version, it failed once, and then worked on the second try. This patch is only adding a comment for the licensing.

Please advise.

Patch-set: 2