Update patch set 2
Patch Set 2:

I fixed the issue that was introduced.  We had more merges yesterday.  I'm going to verify that the new patches didn't introduce any failures in tests when using django 1.6. When I've verified there are no breaking changes I'll update this patch again. 

Django 1.6 is an FFE for horizon in icehouse.

Patch-set: 2