Update patch set 1
Patch Set 1:

because it doesn't actually work. It turns out we never were testing it because of a break in refactoring in devstack.

When we tried to fix it, we were blocked on merging that on this, because Ceilometer doesn't actually support sqla 0.9 (due to some tricky things they do in their migrations).

Patch-set: 1