Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

A good addition, I like it. As the functionality gets added into clients, I think it's worth noting that it only works if you're using bash or zsh. Nothing which needs to be said here, but just good to keep in mind so we can document it as we add it to clients.

Patch-set: 3
Label: Code-Review=+1