Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

(3 inline comments)

I'd like the whitespaces removed.  I think it'd also be good to get a heat and oslo.messaging core in here to review this.

Verified it's on pypi: https://pypi.python.org/pypi/trollius

It's ASL 2.0 licensed.

Looks like it's making it's way through packaging for Fedora/RHEL:

https://bugzilla.redhat.com/show_bug.cgi?id=1066238

And is already packaged in Debian.

From the docs looks like it's handling python 2.6-3.4.

Doing a diff of 0.1.4 and 0.2 setup.py there are no runtime dependency changes so that's good:

https://bitbucket.org/enovance/trollius/diff/setup.py?diff1=429bf62d2636&diff2=d04761ad7717e94155e7e77680ba14884f7edbdf&at=trollius

Patch-set: 1
Label: Code-Review=-1