Update patch set 1
Patch Set 1:

Please get the bug in the commit message as justification for the change, I think that's kind of expected on these changes since it's not only glance that's updated, it's all projects using this library.

And looking at that bug, it looks like a feature hidden in a bug report, and I can't access the glance patch so it's either a draft or something else, but it doesn't sound like something that's going to land in icehouse-rc1.

Patch-set: 1