Update patch set 1
Patch Set 1:

Matt - if your concern is about icehouse-rc1: we can wait for juno to be opened to change the requirement. I am fine with that.
As long as the requirement is updated at the very beginning of Juno, there should not have any problem. If it is not the case, then, Glance development will be blocked by this.

The storage policy support is currently a bug but I agree with you (and had in mind) to mark the bug invalid and to make it a blueprint.

Patch-set: 1