Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

(1 inline comment)

It's fine to get it in, but I don't think that code in oslo.messaging or heat is going to land in icehouse so we should leave this out until the patches that need it are going to move, otherwise deployers would need to pick this up yet in icehouse if it lands and merges.  I guess the heat team could just -2 the sync patch to heat when it shows up.

@Steven @Mark - thoughts?

Patch-set: 2
Reviewer: Gerrit User 6873 <6873@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1