Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

This was used in Neutron for some time without realizing that the code touched is from incubator. This is a 'backport' from neutron to incubator to make the code in both places exactly the same.

Revealed when doing https://review.openstack.org/#/c/80998/

Patch-set: 2