Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

Actually, I've sent a patch to the Debian bug tracker, for the package to support 1.13:

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=697600

If that is applied, and we effectively get 1.13 + Python3 support in the package, then why not?

In the mean time (until the package is fixed), I agree we shouldn't just change the global-requirements.txt. Let's please just leave this changed pending and not merged.

Patch-set: 1
Label: Code-Review=-1