Update patch set 5
Patch Set 5: I would prefer that you didn't merge this

Please add a comment to the new line with the license. See https://wiki.openstack.org/wiki/Requirements for an example.

What is the Python 3 support for this new library? It's not clear from the setup.py or PyPI metadata.

I see that this is a pretty new library. How stable is the API? Should we include an upper-bound on the version range?

Did https://pypi.python.org/pypi/zfec fail to meet our needs for some reason other than the license?

Patch-set: 5
Label: Code-Review=-1
Label: Workflow=0