Update patch set 6
Patch Set 6:

> Please add a comment to the new line with the license. See https://wiki.openstack.org/wiki/Requirements for an example.

Done.  (I didn't see other patches going into requirements tree following this rule so I added the "BSD licensed" comment in the commit message)


> What is the Python 3 support for this new library? It's not clear from the setup.py or PyPI metadata.

The library has been tested to work on Python 2.6, 2.7 (Swift supports only 2.x at the moment which was our starting point).  3.x support is work-in-progress at the moment and should be supported shortly.


> I see that this is a pretty new library. How stable is the API? 

The library was developed with blessing from Swift core developers - here is the community forum used for discussion:  https://trello.com/b/LlvIFIQs/swift-erasure-codes-and-storage-policies.

The API definition came from Kevin Greenan who is a well-known authority on Erasure Codes in the academia and the storage community, with validation by Intel's Storage Division.  

The API is stable at this point and we cannot think of anything that's changing shortly - that said, we will be the ones incorporating it into Swift and the users of the library in Swift.


> Should we include an upper-bound on the version range?

The latest version is the stable version at this point.


> Did https://pypi.python.org/pypi/zfec fail to meet our needs for some reason other than the license?

In addition to the license incompatibility, this library was developed for performance reasons.  PyECLib is simply a Python convenience wrapper on top of multiple C libraries, primarily jerasure and isa-l.  jerasure and isa-l are probably the best performance and most widely-used erasure code libraries.  

Additional features in PyECLib include additional information in headers for Erasure Code fragment integrity checks (CRC/algsig), and also alignment and size information for the encoded fragments which comes handy when dealing with Erasure Codes at the Swift level.

Patch-set: 6
Reviewer: Gerrit User 7485 <7485@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0