Update patch set 2
Patch Set 2:

Hi Doug:

I looked a bit deeper into this, and it looks like quite a few of the trove tests would need to be updated if we had to use WebTest instead of wsgi_intercept. (These tests use the real python-troveclient to test the API through the "fake" monkey-patched trove wsgi-app.) If the wsgi_intercept test-dependency is something that you feel strongly against, we're (trove) willing to take on the task of looking into WebTest and trying to move our tests towards using it but this probably won't land until juno. However, if it's okay with you, I'd prefer accepting this trove dependency since quite a few existing tests trove rely on it.

Thanks!

-Nikhil

Patch-set: 2
Label: Code-Review=0