Add a file that can detail the requirements files
From the ML discussion there was a desire to provide
a more detailed version of the requirements that has
information such as license, author... This script
gathers such information from the pypi json API and
writes it to corresponding files that can be used by
automated programs that want to determine this information
without having to go to pypi themselves.

This can be useful for doing requirement license
auditing for example.

To run:

$ python detail.py global-requirements.txt

This will gather the information about the requirements
from pypi (if possible) and then write a new file located
at global-requirements.json with this information.

Change-Id: I710feb7038eec0c916593cce0837823e7f4fce15