Using happybase 0.8
The only OpenStack package that uses happybase is Ceilometer, and
it works perfectly well with happybase 0.8. The only thing is that
version 0.6 and 0.7 are borken, and the fixes have been made into
version 0.8 (according to the Ceilometer contributors I have talked
with). As Debian upgraded to happybase 0.8 and that everything
worked well with it, the best course of action for Icehouse is to
at least allow 0.8 to be in use. If there's concerns about this,
then we could do: >=0.4,!=0.6,!=0.7, but I think it's best to simply
use version 0.8.

Change-Id: I071f9994c55a343766703f0913c65b63bf26627b