Add junitxml to global requirements
In Rally we are going to make better integration with tempest.
Tempest output is in subunit, which are difficult to parse.

`python-subunit` package provides converter(subunit2junitxml) subunit
output to JUnit XML representation.
XML representation is more friendly for Python and can be parsed
using builtins.

Based on the fact that python-subunit is already in global-requirements,
it would be nice if junitxml will be added there, since it required for
python-subunit.

Change-Id: I7d8d5b50a1d9de7614830f6ede91787b94722e93