Update patch set 3
Patch Set 3: I would prefer that you didn't merge this

The problem is this merge immediately pushes that to all the projects. Which we don't want if they are incompatible with hacking.

Hacking + requirements needs to be atomic. I think that just means hacking needs to be taking out of g-r enforcement, which is fine, because it's not part of the devstack stack.

Patch-set: 3
Label: Workflow=0