Update patch set 4
Patch Set 4: I would prefer that you didn't merge this

I'm currently -1 on this because we already have httpretty (and the 2 look like large overlaps in providing the same function). If this is better than httpretty we should figure out how to get rid of httpretty in projects as part of this.

We should, whenever possible, not make our testing requirements diverge in OpenStack projects, because that makes us have a lot more friction in contributors contributing across projects.

Patch-set: 4
Label: Code-Review=-1