Update patch set 1
Patch Set 1:

Pavlo,

global requirements is used to populate the pypi.openstack.org mirror (which we use for reliability) , so the only way to add a new version to the mirror is through global requirements.

There was a long discussion in the infra channel about the best way forward for this, and this is the second patch in the series. The first has already landed. https://review.openstack.org/#/c/84913/

Patch-set: 1