Update patch set 2
Patch Set 2:

Hmm, the  bug says this:

"Neither it happened on Fedora19/20 or RHEL-6.5 with devstack/tempest master branches."

We run Tempest against Icehouse on RHEL 6.5 and Fedora 19 internally and haven't seen any problems with this.  On RHEL 6.5 we're using sqlalchemy 0.7.9.  On fc19 we use what's in the distro, which is I think 0.8.2.

But if this was a problem for icehouse with rhel 6.5 and sqlalchemy 0.7.9 we should have seen it also.

Patch-set: 2