Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

The bug says it's failing on 0.7.8:

> keystone.common.wsgi File "/usr/lib64/python2.6/site-packages/SQLAlchemy-0.7.8-py2.6-linux-x86_64.egg/sqlalchemy/orm/session.py", line 405, in __exit__

The bug also says it's failing with RDO-Icehouse on RHEL 6.5:

"With RDO-Icehouse (m2 testday packages) on RHEL-6.5,
negative Tempest identity tests fails as Keystone responds with HTTP 400."

But then goes on later to say it's OK with RHEL 6.5:

"Neither it happened on Fedora19/20 or RHEL-6.5 with devstack/tempest master branches."

Keystone's stable/icehouse requirements are sqlalchemy >= 0.7.8:

https://github.com/openstack/keystone/blob/stable/icehouse/requirements.txt#L11

I've had no problems with tempest/master on stable/icehouse with RHEL 6.5 and sqlalchemy 0.7.9.  I haven't tried 0.7.8.

We don't use RDO, we have our own install scripts.  But this sounds like an RDO issue.  From the bug report it's updating sqlalchemy to version 0.9.4 automatically, which falls within the supported versions for keystone on stable/icehouse, but that doesn't mean sqlalchemy-migrate supports that version.  Actually, sqlalchemy-migrate doesn't support sqlalchemy >= 0.9 until migrate version 0.9, so it seems like you maybe also need a newer sqlalchemy-migrate version?  We're running with migrate 0.9 for Icehouse.

Even if you did update the requirements for sqlalchemy to 0.8.1, the bug says this is an Icehouse problem - so are you going to propose the requirements change as a backport to stable/icehouse?  That would have impacts to packagers/deployers for Icehouse.

Patch-set: 2
Reviewer: Gerrit User 6873 <6873@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1