Update patch set 2
Patch Set 2:

Had a chat with sdague on irc, and he suggested that rather than add oauthlib to requirements in stable/havana, to just make the keystoneclient code fail a bit more gracefully. 

I made the changes to https://review.openstack.org/#/c/81980/ but am awaiting feedback. I ask that this patch not get merged until we know definitely which direction we choose to follow.

Patch-set: 2