Update patch set 1
Patch Set 1:

It allows to add the output of a program to a document. I added it to fix a doc issue here: https://review.openstack.org/#/c/45520/

So when the help output of the get changed it automatically updates the doc.
I implemented it based on Clark Boylan's recommendations in the review.

Patch-set: 1