Update patch set 3
Patch Set 3:

Hi,

Well, if we limit netifaces as <=0.8, then it is easy to down-prioritize the importance of figuring out why it is actually failing, but on the other hand we do not have to deal with the gate failures again, so maybe it will be better. I'm still confused about that 0.10.0 on Ubuntu precise works fine, but I can sleep with it.

Best Regards,
Ildiko

Patch-set: 3