Add pypcap to global-requirements
Ironic-python-agent needs a way to obtain LLDP info. As far as I can
tell, the two ways to do this are:
* capture an LLDP packet and parse it
* run lldpd and lldpctl to obtain the info

The latter is a poor choice for IPA, as lldpd requires:
* installing perl
* running a daemon

IPA is baked into a ramdisk and is intended to run as a single
process. So the perl dependency bloats the image, and running a
daemon is undesirable.

Our plan is to capture and parse LLDP packets in python, and as
far as I can tell pypcap is the best tool for the job.

Change-Id: I297a0505b93925d85fb551d895bdb7d79e26970c