Update patch set 2
Patch Set 2: Code-Review+1

Having access to the docker libraries would be useful for both solum and the docker driver ( for when it gets merged back, good to have some of the plumbing available ).

It could also be good for other projects that might want to use docker containers to perform functions for them, or jail a process away from the OS ( like needing a root user to perform a task ) but not actually use the nova docker driver to do so.

Patch-set: 2
Reviewer: Gerrit User 7858 <7858@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1