Add doc8 a documentation style checking package
Many of the projects are starting to do style checks
on there specs files or other rst files so it seems
appropriate to use a small utility library to do these
kinds of checks (to avoid repeated copied code).

Doc8 starts to create a path for a common rst documentation
checking suite that can avoid the 'test_titles.py' tests
that are appearing in the various specification repositories.

Projects that currently can benefit from this package:

- TaskFlow
- Cinder
- Others with spec repos (nova, glance, oslo...)

This is cherry picked from 83b3a598558d1ce5 and is needed
so that libraries that have added doc8 to there testing
requirements can still function correctly (this is after all
just a testing requirement, so this should be harmless).


Change-Id: Ie942497d88ad4dc9a4b3e58bea213354ec5513c5