Add wrapt for taskflow build
This change adds wrapt to the stable/icehouse requirements list
so the check-grenade-dsvm job, which tries to verify that the
requirements for the projects are correct across releases,
will allow the new taskflow to use wrapt as a runtime dependency.
Nothing else currently uses wrapt but taskflow will use it to
provide a correct and unified decorator for usage in its
deprecation module.

Change-Id: I5efdea41bcbe049a2a13742363d83e260f6b814c