Update patch set 1
Patch Set 1:

I think I figured this out.

We're not hitting this with six 1.5.2 because we're using oslo.messaging 1.3.0a9 which didn't have this change:

https://github.com/openstack/oslo.messaging/commit/13608437bb8d491badbc6c906f8d60b51e52cd79

I'm not sure what was being used in the bug reported, i.e. is RDO building from trunk for it's client packages in Icehouse?

Anyway, that's why we didn't have this issue.

Also, the requirement for oslo.messaging in stable/icehouse is this:

oslo.messaging>=1.3.0a9

And 13608437bb8d491badbc6c906f8d60b51e52cd79 is not in the 1.3.0 release of oslo.messaging, which is the latest release, so someone must be building from trunk of oslo.messaging - and this means that oslo.messaging should have it's requirements.txt file updated for six since it's using that feature that's only in >= 1.6.0.

Patch-set: 1