Add dbus library, needed for the DRBD cinder driver integration.
DRBDmanage is a cluster-wide service to provide replicated storage
with a simple administration interface; its daemon listens on
the system dbus for commands.

The DRBD cinder volume driver needs to talk to the daemon, and so
needs the dbus library.

The library is available as python-dbus (deb), and dbus-python (rpm)
in most distributions.

Change-Id: I7c0ca5c58f7b533d6fc93ec52f2dae11f69a3db1