Update patch set 2
Patch Set 2: Code-Review+2

Based on a discussion about this with fungi on IRC, I think we want to go ahead and relax the lower bounds to let requirements changes pass the "broken" test script still present on some of the test nodes. Those are being replaced, but that will take some time. After the fix is distributed, we can look at the lower bound requirement again and decide if we need to raise it or if this range is accurate.

Patch-set: 2
Reviewer: Gerrit User 2472 <2472@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2