revert hacking to 0.8 series
the hacking release came too late in the development cycle to make
it something that's part of this current release. We need to
revert back to the 0.8 series for juno otherwise we're going to
generate a ton of unrelated churn when we need to be solving
real problems in the code.

Change-Id: I717408adaad474db8ce032994cd07855ab93f0b4