Update patch set 1
Patch Set 1:

So I have a function that I would like to get rid of in taskflow that I put upstream in 1.7, I'm sure others could use it as well (along with the other changes that apparently bring improvements).

https://bitbucket.org/gutworth/six/pull-request/32/add-a-wraps-helper/diff was that change, and it got merged into six in the 1.7 series.

Is that enough of a case to bump it, idk, but it doesn't seem like it would hurt anything to do this.

Patch-set: 1
Reviewer: Gerrit User 1297 <1297@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0