Add doc8 for taskflow testing build
This change adds doc8 to the stable/icehouse requirements list
so the check-grenade-dsvm job, which tries to verify that the
requirements for the projects are correct across releases,
will allow taskflow to use doc8 as a testing time dependency.

Nothing else currently uses doc8 but taskflow will use it to
check its own rst formatted documentation against the doc8 provided
style and formatting checks.

Change-Id: I7545bfce0130f31010a34d10663b6d901e507b69