Update patch set 2
Patch Set 2:

This addition of enum34 is not related directly to the current validation framework of Neutron.
It is about the ability to use python 3.4 enum in Neutron.
It is true that the new validation framework will be able to support enums and therefore it makes sense to add this ability.

Patch-set: 2
Reviewer: Gerrit User 6447 <6447@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0