Add osprofiler to global requirements
osprofiler is special designed super tiny library to make a cross
service and project profiler. It means that we will get only 1 trace
per request that goes via different services and projects.

For example if we add osprofiler to cinder and nova we can get trace
like this: http://pavlovic.me/rally/profiler/
(on graphic we are creating volume from image, then boot instance
 from volume and after all delete vm)

As library was designed specialy for OpenStack it is quite easy for
integration with it.

To make it work we need only next patches:

*) 1 patch in Ceilometer (to collect info)
   https://review.openstack.org/#/c/100239/

*) 1 patch to oslo.messaging (for cross service)
   http://goo.gl/nCK21D

*) 1 patch in all python clinet (for cross project)
   http://goo.gl/P0ocrr

*) 1 patch in all project
   http://goo.gl/yY42jd

This code should be always in source tree. Basicaly it does
nothing unit profiler is initalized.

Profiler is initialized only for request that has special
header with signed (by hmac) inforamtion. So it's safe to
keep it turned on in production.

More about lib:
https://github.com/stackforge/osprofiler/blob/master/README.rst

Change-Id: Ia4a2b2f2402ca60f25bf05c40dc07fc80c8e5ad2