Remove solum from the projects list
Ideally we want to be here, but we need to depend on
mistralclient, which is not in the global requirements
(and unlikely to get there until it's integrated afaik).

This will give us access to pypi.python.org but we will
lose the automatic updates to our requiremnts.txt which
is unfortunate.

Change-Id: Ifecff2254a8bc9e83a54869312053fdad8a51c34