Add python-mistralclient to requirements
Solum is adding new features that allow for user customizable CI
pipelines. Our first implementation uses mistral as the workflow
service. Note that stackforge/solum is currently listed in
projects.txt, and would like to remain there for two key reasons:

1) To use the same set of requirements of OpenStack distributions.
2) To benefit form the automated Jenkins bot reviews as requirements
change.

Solum is deeply integrated with a number of OpenStack services, so
it's critically important to our success to be closely aligned on
requirements. We are reluctant to implement our own workflow service
when one already exists in the stackforge community.

Other alternatives were explored. For reference, see:
http://lists.openstack.org/pipermail/openstack-dev/2014-June/038609.html

Change-Id: I6bcf796d9b62aa34bd5017ae6775a050e5c712af