Update patch set 1
Patch Set 1: -Code-Review Workflow-1

We deliberately only upload wheels for pre-releases

I guess it should have been explained fully in I060b59d8adc586ff37065fa4f6172a592ee3f903 but the long and the short of it is that we don't want to expose pre-releases to users who only want the final releases. We can do that with pip 1.4, but pip 1.3 doesn't have that support. Conveniently, 1.3 also doesn't have support for wheels so we ensure that 1.3 users don't see pre-releases simply by not upload sdists of pre-releases

We'll need to remove the --no-use-wheel in order to use pre-releases

Patch-set: 1
Label: Workflow=-1