Update patch set 1
Patch Set 1:

Ah, right. It seems to me we also really shouldn't be depending on a pre-release version in global-requirements. If all of OpenStack is going to get upgraded to this, it's not a pre-release at that point, since our active development relies on stuff being stable and functional.

Patch-set: 1