Update patch set 1
Patch Set 1:

isn't juno pre-release at this point? My understanding is that oslo.config will get a stable release before juno is released.

The best way to ensure stability and functionality is by using it and testing it.

The alternative to use this oslo library is syncing code from oslo-incubator... using prerelease libs seems like a lesser of 2 evils.

Patch-set: 1