Update patch set 1
Patch Set 1:

I guess my question is: does it actually make sense to tie oslo library releases to the overall release schedule? It's fairly confusing to me to put alphas in here (for other libraries, like the client ones, we depend on released versions and they do releases beyond just the lettered ones). Maybe this is outside the scope of this review though :-)

Patch-set: 1