Update patch set 1
Patch Set 1:

> honestly, I don't like this in straight up global 
> requirements, as that means any project can add it. 

What is the downside of this? Is a particular project second class just because it's in Stackforge? I am failing to understand how this is any different than any other project on pypi. Don't think of Mistral as a Stackforge project. Think of it as a pypi library. That's how we are using it.

> Doing this kind of stackforge coupling here means that 
> now solum and mistral are an atomic unit to come forward
> for incubation (should either care to).

No, we only have a dependency on the pypi version of the mistral client. If Mistral wants to file for incubation, that does not mean that Solum needs to.

We need to rely on *some* workflow service. Would you prefer that we use Celery or Spiff workflow? Would you prefer that we copy the python-mistralclient code into our code base?

> I realistically think stackforge projects should be
> removed from projects.txt. 

We are benefiting significantly form running on the same set of requirements as OpenStack. If we were an incubated project, would that mean we would be prohibited from using Mistral because that project is not incubated. Why can we add random pypi projects, and not other Stackforge projects? This seems like a rather serious double standard.

> I'm fine if there is some alternative to g-r for 
> stackforgey things, but I *really* don't like a new 
> requirement being proposed first by a stackforge project.

If other projects use python-mistralclient, why is that bad? How is that any worse than someone asking for spiff workflow?

Thanks.

Patch-set: 1