Update patch set 1
Patch Set 1:

Sean I think brings up two, unrelated issues:

1) Is the question of mistral being in stackforge. I don't think that really means anything, since we can still evalutate it on the same basis as every other PyPI project.

2) Is the fact that Solum is in stackforge, and should we add new things to global-requirements for things in stackforge.

The answer to (2) comes down to "what global-requirements" is for. If it's just a list of any package someones wanted which passes a certain set of tests (license, py3k, maintaince, etc.), then sure, why not add it? On the other hand, if it's supposed to be something more "official", like the reccomended set of libraries or something, I can see the concern.

That said, I think there needs to be /some/ way for stackforge projects to be able to use libraries that aren't on the list, but still get version upgrades for the rest of their dependencies. Either they need to be able to add them to global-requirements.txt, or there needs to be some other process.

Patch-set: 1
Reviewer: Gerrit User 7680 <7680@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0