Update patch set 1
Patch Set 1:

"Is a particular project second class just because it's in Stackforge?"

Yes. That's actually the whole point of stackforge, second class citizens that can take advantage of our infrastructure when it doesn't get in the way with the integrated release. Changing our requirements process feels like getting in the way.

I'm entirely happy to take this to the dev list. I really don't think we should be doing stackforge only things in requirements.

Patch-set: 1