Update patch set 2
Patch Set 2:

@mordred just let me know that they are updating the way the pypi mirroring system is working (have been working on this over the weekend) which should obviate the need for this change. 

In a nutshell, all of pypi will be mirrored, and stackforge projects will be permitted to add additional requirements that are not part of global-requirements.txt while OpenStack project will still have the same restrictions.

This should eliminate the need for us to change our position in the projects.txt file, or add additional requirements to the global-requirements.txt file.

Patch-set: 2