Update patch set 1
Patch Set 1:

Edward - no, that's not the way it works

1.3.0aN is an alpha release of 1.3.0 - it precedes 1.3.0

1.3.0 is the version which was released with icehouse and stable/icehouse should continue to use it ... until the final 1.4.0 release is released

1.4.0.0a2 is an alpha release of 1.4.0.0 - stable/icehouse shouldn't be using it, not because it's a newer version than 1.3.0, but because it's an alpha version

if stable/icehouse does this:

  oslo.messaging>=1.3.0

it will use 1.3.0 and, later when it is released, 1.4.0. It will not, however, use 1.4.0.0aN

right now it does:

  oslo.messaging>=1.3.0a9

and this causes pip to consider *any alpha* versions, not just the 1.3.0aN alpha versions. This is why 1.4.0.0aN is being pulled in.

Patch-set: 1