Update patch set 5
Patch Set 5: Code-Review-1

I'm very worried about the minimum versions which are in this patch. For example, we have:

xstatic-jquery>=1.10.2.1

In Debian, we have libjs-jquery to provide version 1.7.2+dfsg-3 only. When packaging xstatic-jquery, it will follow what the libjs-jquery package also provides, and do a symlink to it. libjs-qunit in Debian is only version 1.14.0 and so on.

So, before approving this, could you please make sure that you really need these versions, and for what reason? I would really appreciate to avoid having to upgrade each and every of these Javascript libraries in Debian.

Typically, javascript library versions are a pain to maintain in a distribution, because they don't have clear API compat scheme. For example, even if a function prototype doesn't change, sometimes the output of the function change from one version to the next. It's very hard to find out for a package maintainer. So I would recommend being extremely careful, and make sure to support older (AND newer) versions if possible.

Also, that's a whole bunch of new dependencies that you're proposing here. Are you sure absolutely all of them are needed? That's a lot of new package to do and maintain, and x2, since there's going to also be the xstatic packages doing symlinks to them. Just adding all of that at once is kind of scary to me, as often, these libraries come with their set of security issues.

Your thoughts welcome! :)

Patch-set: 5
Label: Code-Review=-1