Update patch set 1
Patch Set 1: Code-Review-1

I'm going to soft -1 until at least danpb can weigh in purely based on the fact that 1.2.5 isn't anywhere near available in RHEL 6.5, this is the version I see in the RHEL 6.5 distro:

0.10.2-29.el6_5.9

Looks like RHEL 7 RC2 has version 1.1.1.

The other thing this might push is there has been talk for awhile about removing the fake libvirt stuff from nova's unit tests so that we could be sure we're always running against a real libvirt backend, that might get us closer to that, but then we wouldn't have a way of unit testing experimental features in the libvirt driver in nova for veresions of libvirt that we don't test in the gate...which is another issue in a separate topic.

You might also want to call this out in your commit message for reference:

https://wiki.openstack.org/wiki/LibvirtDistroSupportMatrix

Patch-set: 1
Label: Code-Review=-1