Update patch set 2
Patch Set 2: Code-Review+1

Note that libvirt python versions are not connected to libvirt versions.

ie any libvirt-python version can compile against any version of libvirt from 0.9.11 onwards. It will only ever expose methods + APIs that were present in the version of libvirt it builds against. So in that sense the version of libvirt python isn't really very important. All it is doing is ensuring we have any relevant bug fixes in the python<->C glue layer, it doesn't affect the APIs that the module exposes to us.

So from that POV, we might as well pick the latest libvirt-python version available.

The only thing to be wary of is that our "official" min declared libvirt version for Nova is 0.9.6, which the pypi libvirt-python will not build against. I don't think that's a big issue though - people on such ancient libvirt really would be better served by upgrading, rather than us pretending it is a good platform on which to run devstack and/or gate tests. And people in production with 0.9.6 can still continue with their existing non-pypi libvirt-python install if they really need to.

Patch-set: 2
Label: Code-Review=+1