Update patch set 6
Patch Set 6:

Hi,

Here's my findings, in 3 categories.

1/ Already in Debian and poses no problem:
+xstatic-angular>=1.2.1.1  # MIT License
+xstatic-angular-cookies>=1.2.1.1  # MIT License
+xstatic-angular-mock>=1.2.1.1  # MIT License
+xstatic-d3>=3.1.6.2  # BSD License (3 clause)
+xstatic-rickshaw>=1.4.6.2  # BSD License (prior)

2/ Not in Debian (will need packaging of both libjs-* and python-xstatic-*):
+xstatic-bootstrap-datepicker>=0.0.0.1  # Apache 2.0 License
+xstatic-bootstrap-scss>=2.0.1.1,<3  # Apache 2.0 License
+xstatic-hogan>=2.0.0.2  # Apache 2.0 License
+xstatic-jasmine>=1.3.1.1  # MIT License
+xstatic-jquery-migrate>=1.2.1.1  # MIT License
+xstatic-jquery.bootstrap.wizard>=1.0.0.1  # MIT License
+xstatic-jquery.quicksearch>=2.0.3.1  # MIT License
+xstatic-jsencrypt>=2.0.0.1  # MIT License
+xstatic-spin>=1.2.5.2  # MIT License

3/ Already in Debian in lower version:
+xstatic-jquery.tablesorter>=2.14.5.1  # MIT License
Debian has: 2.0.5b

+xstatic-qunit>=1.14.0.1  # MIT License
Debian has: 1.14.0

So, I'm ok with everything that is in 1/, I'll have to do the packaging for what's in 2/. The problem is for what's in 3/. Could you lower qunit to 1.14.0, and see if we really need jquery.tablesorter>=2.14.5.1? If we could live with 2.0.5b, that'd be great.

Then, as per our IRC chat, jquery >= 1.10.2 is a problem, and not sure how to fix that. It'd be great if Horizon could live with version 1.7.2.

Last: thanks a lot for your patience with me! :)

Cheers,

Thomas

Patch-set: 6